move index.Write, rename Object type constants to include Type and migrate Object tree to objects package
